Add license and authors files
While the upstream GitHub repo does not, at time of writing, contain a
LICENSE file, it was published to addons.mozilla.org under the Mozilla
Public License 2.0. As such, I believe in good faith that adding this
LICNESE file and list of AUTHORS is in keeping with both the spirit and
the letter of the license under which the original browser extension was
published.